Initial login modals introduced
