+++ client/src/components/Login/LoginForm.js
@@ -1,6 +1,8 @@
 import React, { Component } from "react";
 import { Redirect } from "react-router-dom";
 import googleButton from "./google_signin_buttons/web/1x/btn_google_signin_dark_normal_web.png";
+import "./login.css";
+var props;
 
 class LoginForm extends Component {
   constructor(props) {
@@ -28,38 +30,63 @@ class LoginForm extends Component {
       redirectTo: this.props.redirectAfterLogin
     });
   }
-
-  render() {
-    if (this.state.redirectTo) {
-      return <Redirect to={{ pathname: this.state.redirectTo }} />;
-    } else {
-      return (
-        <div className="LoginForm">
-          <h1>Login form</h1>
-          <form>
-            <label htmlFor="username">Username: </label>
-            <input
-              type="text"
-              name="username"
-              value={this.state.username}
-              onChange={this.handleChange}
-            />
-            <label htmlFor="password">Password: </label>
-            <input
-              type="password"
-              name="password"
-              value={this.state.password}
-              onChange={this.handleChange}
-            />
-            <button onClick={this.handleSubmit}>Login</button>
-          </form>
-          <a href="/auth/google">
-            <img src={googleButton} alt="sign into Google Button" />
-          </a>
-        </div>
-      );
+    render() {
+        if (this.state.redirectTo) {
+            return <Redirect to={{ pathname: this.state.redirectTo }} />;
+        } else {
+            return (
+                <div className="modal-wrapper LoginForm"
+                    style={{
+                    transform: props.show ? 'translateY(0vh)' : 'translateY(-0vh)',
+                    opacity: props.show ? '1' : '0'
+                    }}>
+                    <div className="modal-header">
+                        <h1>Sign In</h1>
+                        <span className="close-modal-btn" onClick={props.close}>×</span>
+                    </div>
+                    <div className="modal-body">
+                        <form name="signin" id="modalForm">
+                            <div className="inputField">
+                                <label id="label-username" htmlFor="username">Username:</label>
+                                <input
+                                  className="inputField form-control" 
+                                  id="username-input"
+                                  type="text"
+                                  name="username"
+                                  value={this.state.username}
+                                  onChange={this.handleChange}
+                                  placeholder="Enter username"
+                                  />
+                            </div>
+                            <div className="inputField">
+                                <label id="label-password" htmlFor="password">Password:</label>
+                                <input
+                                    className="inputField form-control" 
+                                    id="password-input"
+                                    type="password"
+                                    name="password"
+                                    value={this.state.password}
+                                    onChange={this.handleChange}
+                                    placeholder="Enter password"
+                                  />
+                            </div>
+                        </form>
+                    </div>
+                    <div> 
+                        <button className='submit-button btn-block' onClick={this.handleSubmit}>Login</button>
+                    </div>
+                    <div>
+                        <a href="/auth/google">
+                            <img src={googleButton} alt="sign into Google Button" />
+                        </a>
+                    </div>
+                    <div className="modal-footer">
+                        <a href="/signup" className='footerText'c>Create Account</a>
+                    </div> 
+                </div>
+            );
+        }
     }
-  }
 }
 
 export default LoginForm;
